Rearrange serialization files
This breaks out the serialization stuff from one file to some smaller ones
and fixes the corresponding import / naming issues
